rwd pt 2

+++ src/components/cards/AlbumCard.tsx
@@ -1,6 +1,5 @@
 import NextLink from 'next/link';
 import { Box, Center, Link } from '@chakra-ui/react';
-// import Box from 'src/elements/text/Box';
 import ImgAlbum from '../../elements/images/ImgAlbum';
 
 interface AlbumCardProps {
@@ -50,7 +49,7 @@ const AlbumCard = ({ artist, artistId, name, thumbnail }: AlbumCardProps) => {
 						<Link>{`${short(name)} by ${artist}`}</Link>
 					</NextLink>
 				) : (
-					<Box> {short(name)}</Box>
+					<Box> {name}</Box>
 				)}
 			</Center>
 		</Box>
